Managed to make the app to navigate directly to specific build
