--- src/App.js
@@ -1,108 +1,0 @@
-import React, { Component } from 'react';
-// import logo from './logo.svg';
-import loading from './Images/loading.gif'
-import SuitePane from './SuitePane';
-import SuiteSelectorPane from './SuiteSelector';
-import './App.css';
-
-
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      suites: null,
-      suite: null,
-      buildSelection: null,
-      additional_info: null,
-      filter: 'feature'
-    };
-    this.onSelect = this.onSelect.bind(this);
-    this.onFilterChange = this.onFilterChange.bind(this);
-  }
-
-  componentDidMount() {
-
-    const regexp = /suite\/(\w*)\/build\/(\d*)/;
-
-    if (regexp.test(window.location.href)){
-      var found = regexp.exec(window.location.href);
-
-      this.props.database.ref(`/${found[1]}/${found[2]}`).on('value', (snapshot) => {
-        this.setState ({
-          suites: {loading: 'loading'},
-          suite: snapshot.val().executions || {},
-          additional_info: snapshot.val().additional_info
-        });
-      });
-
-      this.props.database.ref('/').once('value', (snapshot) => {
-        this.setState ({suites: snapshot.val()});
-      });
-    } else {
-      this.props.database.ref('/').once('value', (snapshot) => {
-        this.setState ({suites: snapshot.val()});
-      });
-    }
-  }
-
-  onFilterChange(value) {
-    this.setState({filter: value})
-  }
-
-  onSelect(buildSelection) {
-    if (this.state.buildSelection) {
-      this.props.database.ref(`/${this.state.buildSelection.selectedSuite}/${this.state.buildSelection.selectedBuild}/executions`).off()
-    }
-
-    location.replace(`/suite/${buildSelection.selectedSuite}/build/${buildSelection.selectedBuild}`);
-
-    this.props.database.ref(`/${buildSelection.selectedSuite}/${buildSelection.selectedBuild}/executions`).on('value', (snapshot) => {
-      this.setState ({
-        suite: snapshot.val() || {},
-        buildSelection: buildSelection
-      });
-    });
-
-    this.props.database.ref(`/${buildSelection.selectedSuite}/${buildSelection.selectedBuild}/additional_info`).once('value', (snapshot) => {
-      this.setState ({
-        additional_info: snapshot.val()
-      });
-    });
-  }
-
-  render() {
-    if(!this.state.suites){
-      return(
-        <div className="no-suites-found">
-          <img src={loading} alt=""/>
-          <h1>Looking for saved executions</h1>
-        </div>
-      );
-    } else {
-      if (!this.state.suite) {
-        return(
-            <div className='App'>
-              <div className="no-suite-loaded">
-                <h1>Select suite and build</h1>
-                <SuiteSelectorPane suites={this.state.suites} onSelect={this.onSelect} />
-              </div>
-            </div>
-        );
-      } else {
-        return (
-            <div className='App'>
-              <SuiteSelectorPane suites={this.state.suites} onSelect={this.onSelect} />
-              <SuitePane
-                  suite={{suite: this.state.suite}}
-                  additionalInfo={this.state.additional_info}
-                  onFilterChange={this.onFilterChange}
-                  filter={this.state.filter}
-              />
-            </div>
-        );
-      }
-    }
-  }
-}
-
-export default App;
