+++ src/Modules/SuitePaneRoute/SuitePane/SuiteHeaderPane/index.js
